+++ CHANGELOG.md
@@ -10,9 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Allow to provide JSON response as python values.
 - Mock async httpx requests as well.
+- Allow to provide files and boundary for multipart response.
+- Allow to provide data as a dictionary for multipart response.
 
 ### Changed
 - method can now be provided even if not entirely upper cased.
+- content parameter renamed into data.
 
 ## [0.0.2] - 2020-02-06
 ### Added